feat: Add PDFFilterModel for filtering PDF files and directories with regex support
